init commit
In this first commit, the README.md has great progress already. I was
not sure if I would upload this project to Github, but I decided I would
as I upload even my worst of code.

What is done so far: the base class and associated methods with editing
defaults as well as the base methods for the first conversions;
converting a list to a extreme pretty print string. In addition, the
base PrettyPrinter class is using the pprint module (as stated in the
readme) for classic pretty printing.

Lastly, I would like this to be a package on PyPi one day, so the
setup.py file is also found in this first commit.