Program progress, README update
Added new default: capitalize. See readme for details.

Attempted to add (will continue to try in future) a method to iterate
through the entire input structure and group each of the nested
structures. Making this will simplify the prettify methods as well as
enable higher depth nested structures of different types to be handled.

Now added basic methods for dictionaries, prettify_dict() and
iterate_through_dict().

Lastly, peachy_print() was added. This unifies the pretty print methods
so that the user only has to call peachy_print and it will recognise
which method to use to prettify the structure.