Fixed end newline bug, fixed spacing issue lists
There was a bug where new lines were added to the end of the string
returned. In addition, when there were nested data types of different
kinds, multiple strings were either printed or returned instead of one.
Both of these issues were fixed by created a variable to store the
overall output and finally print it when the initial data structure has
been processed. That way, I can use str.strip() on the final output to
take out the extra new lines.

In addition, a problem arose at some point that I never noticed; when
lists are nested in lists, new lines are added between the lists,
messing up the counter. This was fixed by adding a variable to keep
track of the initial input type. Using this data, I was able to add
logic preventing the addition of a new line between the lists.